+++ lib/src/header_parser/type_extractor/extractor.dart
@@ -15,7 +15,7 @@ import '../translation_unit_parser.dart';
 import '../type_extractor/cxtypekindmap.dart';
 import '../utils.dart';
 
-var _logger = Logger('ffigen.header_parser.extractor');
+final _logger = Logger('ffigen.header_parser.extractor');
 const _padding = '  ';
 
 /// Converts cxtype to a typestring code_generator can accept.